\name{Rosnow}
\alias{Rosnow}
\docType{data}
\title{Rosnow's and Rosenthal's Baseball Performance Data}
\description{
Contrived data set of an hypothetical study about the efficacy of a new training method for improving the performance of baseball players.  2x2 factorial design with two factors: training method and expertise of the player.
}
\usage{Rosnow}
\format{
A  data frame with 36 observations and three columns:
\describe{
\item{\code{method}}{Training method. Factor with levels \code{control} and \code{Ralphing}.}
\item{\code{experience}}{Experience of the player in championship competition. Factor with levels \code{no}, \code{yes}.}
\item{\code{hits}}{Performance scores, measured as total number of hits in an experimentally devised game.}
}
}
\source{
Rosnow, R. L., & Rosenthal, R. (1989). \sQuote{Definition and Interpretation of Interaction Effects}, \emph{Psychological Bulletin}, 105(1), 143-146.
}
\keyword{datasets}
